requirements.txt
joblib==1.3.2
keras==3.1.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
numpy==1.26.4
pillow==10.2.0
pip==24.0
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.2.0
tensorflow==2.16.1